requirements.txt
django>=5.2
twilio>=8.0
python-dotenv==1.2.1
redis>=5.0.0
python-dateutil>=2.8.0